+++ code/sys.ui/react.dev/src/types.ts
@@ -1,3 +1,3 @@
-export type * from './t.def';
+export type * from './t';
 export type * from './ui/Harness/t';
 export type * from './ui/List.Module/t';
